requirements.txt
Django
gunicorn
django-heroku
django-allauth
psycopg2-binary
dj-database-url
channels